+++ internal/provider/resource_fmc_group_policy_test.go
@@ -34,29 +34,16 @@ func TestAccFmcGroupPolicy(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "name", "my_group_policy"))
 	checks = append(checks, resource.TestCheckResourceAttrSet("fmc_group_policy.test", "type"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "description", "My Group Policy object"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "vpn_protocol_ssl", "true"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "vpn_protocol_ipsec_ikev2", "true"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "ipv4_address_pools.0.id", ""))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "protocol_ssl", "true"))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "protocol_ipsec_ikev2", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "banner", "Welcome to the VPN Connection."))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "primary_dns_server_host_id", ""))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "secondary_dns_server_host_id", ""))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "primary_wins_server_host_id", ""))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "secondary_wins_server_host_id", ""))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "dhcp_network_scope_network_object_id", ""))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "default_domain", "example.com"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "ipv4_split_tunnel_policy", "TUNNEL_ALL"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "ipv6_split_tunnel_policy", "TUNNEL_ALL"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "split_tunnel_acl_id", "12345678-1234-1234-1234-123456"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "split_tunnel_acl_type", "ExtendedAccessList"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "dns_request_split_tunnel_policy", "USE_SPLIT_TUNNEL_SETTING"))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "dns_request_split_tunnel_policy", "TUNNEL_SPECIFIED_DOMAINS"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "split_dns_domain_list", "example.com,example.org"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "secure_client_profile_id", "12345678-1234-1234-1234-123456"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "secure_client_management_profile_id", "12345678-1234-1234-1234-123456"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "secure_client_modules.0.type", ""))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "secure_client_modules.0.profile_id", "12345678-1234-1234-1234-123456"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "secure_client_modules.0.download_module", "true"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "ssl_compression", ""))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "dtls_compression", ""))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "ssl_compression", "DISABLED"))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "dtls_compression", "DISABLED"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "mtu_size", "1406"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "ignore_df_bit", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "keep_alive_messages", "true"))
@@ -69,26 +56,21 @@ func TestAccFmcGroupPolicy(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "ssl_rekey", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "ssl_rekey_method", "NEW_TUNNEL"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "ssl_rekey_interval", "60"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "client_firewall_private_network_rules_acl_id", "12345678-1234-1234-1234-123456"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "client_firewall_public_network_rules_acl_id", "12345678-1234-1234-1234-123456"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "secure_client_custom_attributes.0.id", "12345678-1234-1234-1234-123456"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "traffic_filter_acl_id", "12345678-1234-1234-1234-123456"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "restrict_vpn_to_vlan", "100"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "access_hours_time_range_id", "12345678-1234-1234-1234-123456"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "simultaneous_logins_per_user", "3"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "maximum_connection_time", "3600"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "maximum_connection_time_alert_interval", "1"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "vpn_idle_timeout", "3600"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "vpn_idle_timeout_alert_interval", "1"))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "idle_timeout", "3600"))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_group_policy.test", "idle_timeout_alert_interval", "1"))
 
 	var steps []resource.TestStep
 	if os.Getenv("SKIP_MINIMUM_TEST") == "" {
 		steps = append(steps, resource.TestStep{
-			Config: testAccFmcGroupPolicyConfig_minimum(),
+			Config: testAccFmcGroupPolicyPrerequisitesConfig + testAccFmcGroupPolicyConfig_minimum(),
 		})
 	}
 	steps = append(steps, resource.TestStep{
-		Config: testAccFmcGroupPolicyConfig_all(),
+		Config: testAccFmcGroupPolicyPrerequisitesConfig + testAccFmcGroupPolicyConfig_all(),
 		Check:  resource.ComposeTestCheckFunc(checks...),
 	})
 	steps = append(steps, resource.TestStep{
@@ -107,6 +89,30 @@ func TestAccFmcGroupPolicy(t *testing.T) {
 // End of section. //template:end testAcc
 
 // Section below is generated&owned by "gen/generator.go". //template:begin testPrerequisites
+
+const testAccFmcGroupPolicyPrerequisitesConfig = `
+resource "fmc_ipv4_address_pool" "test" {
+  name    = "group_policy_ipv4_address_pool1"
+  range   = "10.10.10.0-10.10.10.240"
+  netmask = "255.255.255.0"
+}
+
+resource "fmc_hosts" "test" {
+  items = {
+    "group_policy_host_1" = { ip = "192.168.10.1" }
+    "group_policy_host_2" = { ip = "192.168.10.2" }
+    "group_policy_host_3" = { ip = "192.168.10.3" }
+    "group_policy_host_4" = { ip = "192.168.10.4" }
+  }
+}
+
+resource "fmc_networks" "test" {
+  items = {
+    "group_policy_network_1" = { prefix = "192.168.20.0/24" }
+  }
+}
+`
+
 // End of section. //template:end testPrerequisites
 
 // Section below is generated&owned by "gen/generator.go". //template:begin testAccConfigMinimal
@@ -126,33 +132,24 @@ func testAccFmcGroupPolicyConfig_all() string {
 	config := `resource "fmc_group_policy" "test" {` + "\n"
 	config += `	name = "my_group_policy"` + "\n"
 	config += `	description = "My Group Policy object"` + "\n"
-	config += `	vpn_protocol_ssl = true` + "\n"
-	config += `	vpn_protocol_ipsec_ikev2 = true` + "\n"
+	config += `	protocol_ssl = true` + "\n"
+	config += `	protocol_ipsec_ikev2 = true` + "\n"
 	config += `	ipv4_address_pools = [{` + "\n"
-	config += `		id = ""` + "\n"
+	config += `		id = fmc_ipv4_address_pool.test.id` + "\n"
 	config += `	}]` + "\n"
 	config += `	banner = "Welcome to the VPN Connection."` + "\n"
-	config += `	primary_dns_server_host_id = ""` + "\n"
-	config += `	secondary_dns_server_host_id = ""` + "\n"
-	config += `	primary_wins_server_host_id = ""` + "\n"
-	config += `	secondary_wins_server_host_id = ""` + "\n"
-	config += `	dhcp_network_scope_network_object_id = ""` + "\n"
+	config += `	primary_dns_server_host_id = fmc_hosts.test.items["group_policy_host_1"].id` + "\n"
+	config += `	secondary_dns_server_host_id = fmc_hosts.test.items["group_policy_host_2"].id` + "\n"
+	config += `	primary_wins_server_host_id = fmc_hosts.test.items["group_policy_host_3"].id` + "\n"
+	config += `	secondary_wins_server_host_id = fmc_hosts.test.items["group_policy_host_4"].id` + "\n"
+	config += `	dhcp_network_scope_network_object_id = fmc_networks.test.items["group_policy_network_1"].id` + "\n"
 	config += `	default_domain = "example.com"` + "\n"
 	config += `	ipv4_split_tunnel_policy = "TUNNEL_ALL"` + "\n"
 	config += `	ipv6_split_tunnel_policy = "TUNNEL_ALL"` + "\n"
-	config += `	split_tunnel_acl_id = "12345678-1234-1234-1234-123456"` + "\n"
-	config += `	split_tunnel_acl_type = "ExtendedAccessList"` + "\n"
-	config += `	dns_request_split_tunnel_policy = "USE_SPLIT_TUNNEL_SETTING"` + "\n"
+	config += `	dns_request_split_tunnel_policy = "TUNNEL_SPECIFIED_DOMAINS"` + "\n"
 	config += `	split_dns_domain_list = "example.com,example.org"` + "\n"
-	config += `	secure_client_profile_id = "12345678-1234-1234-1234-123456"` + "\n"
-	config += `	secure_client_management_profile_id = "12345678-1234-1234-1234-123456"` + "\n"
-	config += `	secure_client_modules = [{` + "\n"
-	config += `		type = ""` + "\n"
-	config += `		profile_id = "12345678-1234-1234-1234-123456"` + "\n"
-	config += `		download_module = true` + "\n"
-	config += `	}]` + "\n"
-	config += `	ssl_compression = ""` + "\n"
-	config += `	dtls_compression = ""` + "\n"
+	config += `	ssl_compression = "DISABLED"` + "\n"
+	config += `	dtls_compression = "DISABLED"` + "\n"
 	config += `	mtu_size = 1406` + "\n"
 	config += `	ignore_df_bit = true` + "\n"
 	config += `	keep_alive_messages = true` + "\n"
@@ -165,19 +162,12 @@ func testAccFmcGroupPolicyConfig_all() string {
 	config += `	ssl_rekey = true` + "\n"
 	config += `	ssl_rekey_method = "NEW_TUNNEL"` + "\n"
 	config += `	ssl_rekey_interval = 60` + "\n"
-	config += `	client_firewall_private_network_rules_acl_id = "12345678-1234-1234-1234-123456"` + "\n"
-	config += `	client_firewall_public_network_rules_acl_id = "12345678-1234-1234-1234-123456"` + "\n"
-	config += `	secure_client_custom_attributes = [{` + "\n"
-	config += `		id = "12345678-1234-1234-1234-123456"` + "\n"
-	config += `	}]` + "\n"
-	config += `	traffic_filter_acl_id = "12345678-1234-1234-1234-123456"` + "\n"
 	config += `	restrict_vpn_to_vlan = 100` + "\n"
-	config += `	access_hours_time_range_id = "12345678-1234-1234-1234-123456"` + "\n"
 	config += `	simultaneous_logins_per_user = 3` + "\n"
 	config += `	maximum_connection_time = 3600` + "\n"
 	config += `	maximum_connection_time_alert_interval = 1` + "\n"
-	config += `	vpn_idle_timeout = 3600` + "\n"
-	config += `	vpn_idle_timeout_alert_interval = 1` + "\n"
+	config += `	idle_timeout = 3600` + "\n"
+	config += `	idle_timeout_alert_interval = 1` + "\n"
 	config += `}` + "\n"
 	return config
 }