+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - (Enhancement) Add support for `fmc_service_access` resource and data source
 - (Enhancement) Add support for `fmc_secure_client_*` resources and data sources
 - (Enhancement) Add support for `fmc_radius_server_group` resource and data source
+- (Enhancement) Add support for `fmc_group_policy` resource and data source
 
 ## 2.0.0-rc4